other admin and super admin

+++ routes/mongoro/Admin/super/category/category.js
@@ -1,6 +1,8 @@
 const express = require('express')
 const router = express.Router()
 const MongoroUserModel = require("../../../../../models/mongoro/auth/mongoroUser_md")
+const OtherModel = require('../../../../../models/mongoro/admin/other/otherAdmi_md')
+const SuperModel = require('../../../../../models/mongoro/admin/super_admin/super_md')
 const CategoryModel = require("../../../../../models/mongoro/admin/super_admin/category/category")
 const dotenv = require("dotenv")
 dotenv.config()
@@ -10,20 +12,31 @@ dotenv.config()
 router.post('/create', async (req, res) => {
 
     try {
-        if (!req.body.name ) return res.status(402).json({ msg: 'please check the fields ?' })
 
-        const validate = await CategoryModel.findOne({ name: req.body.name })
-        if (validate) return res.status(404).json({ msg: 'There is another category with that name ' })
+        if (!req.body.category) return res.status(402).json({ msg: 'please check the fields ?' })
 
-        let category = await new CategoryModel(req.body)
+        const user = await SuperModel.findOne({ email: req.body.super_email })
 
-        await category.save().then(category => {
-            return res.status(200).json({
-                msg: 'Category created successfully ',
-                category: category,
-                status: 200
+        const validate = await CategoryModel.findOne({ category: req.body.category })
+        if (validate) return res.status(404).json({ msg: 'Category exist already ' })
+
+        if (user) {
+            let category = await new CategoryModel(req.body)
+
+            await category.save().then(category => {
+                return res.status(200).json({
+                    msg: 'Category created successfully ',
+                    category: category,
+                    status: 200
+                })
             })
-        })
+
+        } else {
+            res.status(400).json({
+                msg: 'you dont have access to create category',
+                status: 400
+            })
+        }
 
     } catch (error) {
         res.status(500).json({
@@ -33,7 +46,7 @@ router.post('/create', async (req, res) => {
     }
 })
 
-router.get("/all", async (req, res) => {
+router.get("/allcategory", async (req, res) => {
     try {
         const category = await CategoryModel.find();
         res.status(200).json(category.reverse());
@@ -45,11 +58,24 @@ router.get("/all", async (req, res) => {
     }
 })
 
-router.get("/of/:id", async (req, res) => {
+router.get("/allinvited", async (req, res) => {
     try {
-        const user = await MongoroUserModel.find({ category: req.params.id });
+        const category = await OtherModel.find();
+        res.status(200).json(category.reverse());
+    } catch (err) {
+        res.status(500).json({
+            msg: 'there is an unknown error sorry ',
+            status: 500
+        })
+    }
+})
 
-        res.status(200).json(user);
+router.get("/of/:category", async (req, res) => {
+    try {
+
+        const admin = await OtherModel.find({ category: req.params.category });
+
+        res.status(200).json(admin);
     } catch (err) {
         res.status(500).json({
             msg: 'there is an unknown error sorry !',
@@ -63,20 +89,37 @@ router.put('/edit', async (req, res) => {
     let { id } = body;
 
     try {
-        if (!req.body.id) return res.status(402).json({ msg: 'provide the id ?',status: 402 })
 
-        await CategoryModel.updateOne({ _id: id }, body).then(async () => {
-           
-            let category = await CategoryModel.findOne({ _id: id })
-            await CategoryModel.updateOne({ updated_at: category.updated_at }, { $set: { updated_at: Date.now() } })
-            return res.status(200).json({
-                msg: 'Category Updated Successfully ',
-                category: category,
-                status: 200
+        const validate = await CategoryModel.findOne({ category: req.body.category })
+        if (validate) return res.status(400).json({ msg: 'Category exist already ' })
+
+        const check = await CategoryModel.findOne({ _id: req.body.id })
+        if (!check) return res.status(400).json({ msg: 'No category with the provided id ' })
+
+        const user = await SuperModel.findOne({ email: req.body.super_email })
+
+        if (!req.body.id) return res.status(402).json({ msg: 'provide the id ?', status: 402 })
+
+        if (user) {
+
+            await CategoryModel.updateOne({ _id: id }, body).then(async () => {
+
+                let category = await CategoryModel.findOne({ _id: id })
+                await CategoryModel.updateOne({ updated_at: category.updated_at }, { $set: { updated_at: Date.now() } })
+                return res.status(200).json({
+                    msg: 'Category Updated Successfully ',
+                    category: category,
+                    status: 200
+                })
+            }).catch((err) => {
+                res.send(err)
             })
-        }).catch((err) => {
-            res.send(err)
-        })
+        } else {
+            res.status(400).json({
+                msg: 'you dont have access to edit category',
+                status: 400
+            })
+        }
 
     } catch (error) {
         res.status(500).json({
@@ -89,10 +132,24 @@ router.put('/edit', async (req, res) => {
 
 router.delete("/delete", async (req, res) => {
     try {
-        if (!req.body.id ) return res.status(402).json({ msg: 'provide the id ?' })
 
-        await CategoryModel.deleteOne({ _id: req.body.id })
-        res.status(200).json({msg: "Category deleted....",status: 200});
+        if (!req.body.id) return res.status(402).json({ msg: 'provide the id ?' })
+
+        const check = await CategoryModel.findOne({ _id: req.body.id })
+        if (!check) return res.status(400).json({ msg: 'No category with the provided id ' })
+
+        const supers = await SuperModel.findOne({ email: req.body.super_email })
+
+        if (supers) {
+            await CategoryModel.deleteOne({ _id: req.body.id })
+            res.status(200).json({ msg: "Category deleted....", status: 200 });
+        } else {
+            res.status(400).json({
+                msg: 'you dont have access to delete category',
+                status: 400
+            })
+        }
+
     } catch (error) {
         res.status(500).json({
             msg: 'there is an unknown error sorry ',
@@ -105,35 +162,69 @@ router.delete("/delete", async (req, res) => {
 //INVITE    
 router.post("/invite", async (req, res) => {
 
+    const supers = await SuperModel.findOne({ email: req.body.super_email })
+
     const user = await MongoroUserModel.findOne({ email: req.body.email });
 
     const validate = await CategoryModel.findOne({ name: req.body.category })
     if (!validate) return res.status(404).json({ msg: 'There is no such category !' })
 
-    if (user == null) {
-        console.log("User does not exists");
-        res.status(401).json({msg: "sorry..... User does not exists",status: 401});
+    if (!supers) {
+        res.status(400).json({
+            msg: 'you dont have access to create category',
+            status: 400
+        })
+    } else if (user) {
+        res.status(401).json({ msg: "sorry..... This email address alreday exist as a user, Can't be used to register as Admin", status: 401 });
     } else {
-        await MongoroUserModel.updateOne({ _id: user._id }, { $set: { category: req.body.category } })
+        let category = await new OtherModel(req.body)
+        await category.save()
 
-        res.status(200).json({ msg: 'User Invited successfuly !',status: 200 });
+        let transporter = nodemailer.createTransport({
+            service: "hotmail",
+            auth: {
+                user: 'sales@reeflimited.com',
+                pass: 'cmcxsbpkqvkgpwmk'
+            }
+        });
+
+        let mailOptions = {
+            from: 'sales@reeflimited.com',
+            to: req.body.email,
+            subject: 'Invitation',
+            html: `<center><p>You are invited to be an Admin from Mongoro, Input your email and create password </p></center>`
+        };
+
+        transporter.sendMail(mailOptions, function (error, info) {
+            if (error) {
+                console.log(error);
+            } else {
+                console.log('Email sent: ' + info.response);
+            }
+        });
+
+        res.status(200).json({ msg: 'Invited successfuly !', status: 200 });
     }
 
 })
 
 router.post("/disable", async (req, res) => {
 
-    const user = await MongoroUserModel.findOne({ _id: req.body.id });
+    const supers = await SuperModel.findOne({ email: req.body.super_email })
+    const user = await OtherModel.findOne({ email: req.body.email });
 
-    if (!req.body.id) return res.status(402).json({ msg: 'provide the id ?', status: 402 })
+    if (!req.body.email) return res.status(402).json({ msg: 'provide the email ?', status: 402 })
 
-    if (user == null) {
-        console.log("User does not exists");
-        res.status(401).json({msg: "wrong Email !",status: 401});
+    if (!user) {
+        res.status(401).json({ msg: "wrong Email !", status: 401 });
+    } else if (!supers) {
+        res.status(400).json({
+            msg: 'you dont have access ',
+            status: 400
+        })
     } else {
-        await MongoroUserModel.updateOne({ _id: user._id }, { $set: { category: "none" } })
-
-        res.status(200).json({ msg: 'User Disabled successfuly ',status: 500 });
+        await OtherModel.deleteOne({ email: req.body.email })
+        res.status(200).json({ msg: 'Disabled successfully ', status: 200 });
     }
 
 })
